requirements.txt
pyyaml
prettyconf
